I've refactored `backend/main.rs` into multiple modules.
I separated entities (Place, Accommodation, Restaurant, TravelPlan, PlanItem) and their corresponding models, API handlers, and database logic into distinct modules.

- I created `places.rs`, `accommodations.rs`, `restaurants.rs`, `travel_plans.rs`, `search.rs`, and `db.rs`.
- I moved relevant structs and functions from `main.rs` to these new modules.
- I updated `main.rs` to use these modules, including adjustments to the test suite.
- I ensured all necessary items are public and that the project compiles.
- All tests pass after refactoring.
